refactor: modularize core library
